+++ app/src/main/java/fr/eni/ecole/enishop/ui/screens/ArticleDetailsScreen.kt
@@ -1,5 +1,9 @@
 package fr.eni.ecole.enishop.ui.screens
 
+import android.content.Intent
+import android.net.Uri
+import android.util.Log
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -13,56 +17,39 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
 import coil.compose.AsyncImage
 import fr.eni.ecole.enishop.bo.Article
 import fr.eni.ecole.enishop.utils.toFrenchFormat
 import fr.eni.ecole.enishop.utils.toPriceFormat
-import java.text.SimpleDateFormat
-import java.util.Date
+import fr.eni.ecole.enishop.vm.ArticleDetailsViewModel
 
 @Composable
 fun ArticleDetailsScreen(
     articleId: Long,
-    modifier: Modifier = Modifier
+    modifier: Modifier = Modifier,
+    viewModel: ArticleDetailsViewModel = viewModel(factory = ArticleDetailsViewModel.Factory)
 ) {
-    val articlesInMemory = mutableListOf(
-        Article(
-            id = 1,
-            name = "Fjallraven - Foldsack No. 1 Backpack, Fits 15 Laptops",
-            description = "Your perfect pack for everyday use and walks in the forest. Stash your laptop (up to 15 inches) in the padded sleeve, your everyday",
-            price = 109.95,
-            urlImage = "https://fakestoreapi.com/img/81fPKd-2AYL._AC_SL1500_t.png",
-            category = "jewelery",
-            date = Date(),
-        ),
-        Article(
-            id = 2,
-            name = "Mens Casual Premium Slim Fit T-Shirts",
-            description = "Slim-fitting style, contrast raglan long sleeve, three-button henley placket, light weight & soft fabric for breathable and comfortable wearing. And Solid stitched shirts with round neck made for durability and a great fit for casual fashion wear and diehard baseball fans. The Henley style round neckline includes a three-button placket.",
-            price = 22.3,
-            urlImage = "https://fakestoreapi.com/img/71-3HjGNDUL._AC_SY879._SX._UX._SY._UY_t.png",
-            category = "men's clothing",
-            date = Date(),
-        ),
-        Article(
-            id = 3,
-            name = "Mens Cotton Jacket",
-            description = "Great outerwear jackets for Spring/Autumn/Winter, suitable for many occasions, such as working, hiking, camping, mountain/rock climbing, cycling, traveling or other outdoors. Good gift choice for you or your family member. A warm hearted love to Father, husband or son in this thanksgiving or Christmas Day.",
-            price = 55.99,
-            urlImage = "https://fakestoreapi.com/img/71li-ujtlUL._AC_UX679_t.png",
-            category = "men's clothing",
-            date = Date(),
-        )
-    )
+    val article by viewModel.article.collectAsState()
 
-    val article: Article = articlesInMemory.find { it.id == articleId } as Article;
+    // Charge l’article au démarrage
+    LaunchedEffect(articleId) {
+        viewModel.loadArticle(articleId)
+    }
 
-    ArticleDetails(article, modifier)
-
+    if (article == null) {
+        Text("Chargement...", Modifier.padding(16.dp))
+    } else {
+        ArticleDetails(article!!, modifier)
+    }
 }
 
 @Composable
@@ -70,6 +57,9 @@ fun ArticleDetails(
     article: Article,
     modifier: Modifier = Modifier
 ) {
+    val context = LocalContext.current
+    val googleUrl = "https://www.google.com/search?q="
+
     Column(
         verticalArrangement = Arrangement.SpaceEvenly,
         modifier = modifier.fillMaxSize().padding(16.dp)
@@ -78,7 +68,15 @@ fun ArticleDetails(
             text = article.name,
             style = MaterialTheme.typography.titleLarge,
             textAlign = TextAlign.Center,
-            fontWeight = FontWeight.Bold
+            fontWeight = FontWeight.Bold,
+            modifier = Modifier.clickable {
+                val query = "${article.name} EniShop"
+                val intent = Intent(
+                    Intent.ACTION_VIEW,
+                    Uri.parse("$googleUrl$query")
+                )
+                context.startActivity(intent)
+            }
         )
         Surface(
             color = MaterialTheme.colorScheme.inverseOnSurface,
