+++ app/src/main/java/fr/eni/ecole/enishop/vm/ArticleDetailsViewModel.kt
@@ -1,15 +1,17 @@
 package fr.eni.ecole.enishop.vm
 
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
+import androidx.lifecycle.viewmodel.CreationExtras
 import fr.eni.ecole.enishop.bo.Article
 import fr.eni.ecole.enishop.repository.ArticleRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
-class ArticleDetailViewModel(
-    private val repository: ArticleRepository = ArticleRepository()
+class ArticleDetailsViewModel(
+    private val repository: ArticleRepository
 ) : ViewModel() {
 
     private val _article = MutableStateFlow<Article?>(null)
@@ -19,6 +21,16 @@ class ArticleDetailViewModel(
         viewModelScope.launch {
             val result = repository.getArticle(id)
             _article.value = result
+        }
+    }
+
+    companion object {
+        val Factory: ViewModelProvider.Factory = object : ViewModelProvider.Factory {
+            @Suppress("UNCHECKED_CAST")
+            override fun <T : ViewModel> create(modelClass: Class<T>, extras: CreationExtras): T {
+                val repository = ArticleRepository()
+                return ArticleDetailsViewModel(repository) as T
+            }
         }
     }
 }
